// src/app/api/models/receivables.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Aged } from '../models/aged';
import { ReceivableRow } from '../models/receivable-row';
import { ReceivablesSummary } from '../models/receivables-summary';
export interface Receivables {
  as_at: string;

  /**
   * Most owed first.
   */
  rows: Array<ReceivableRow>;

  /**
   * Everyone who owes, whatever the filter.
   */
  summary: ReceivablesSummary;

  /**
   * The rows listed, added up.
   */
  totals: Aged;
}